sessao14/requirements.txt
crewai>=0.28.0
langchain>=0.0.335
openai>=1.3.7
pydantic>=2.4.2
python-dotenv>=1.0.0
